finalize 47.1

+++ dom/indexedDB/test/helpers.js
@@ -113,7 +113,14 @@ function testHarnessSteps() {
 
     let worker = new Worker(workerScriptURL);
 
+    worker._expectingUncaughtException = false;
     worker.onerror = function(event) {
+      if (worker._expectingUncaughtException) {
+        ok(true, "Worker had an expected error: " + event.message);
+        worker._expectingUncaughtException = false;
+        event.preventDefault();
+        return;
+      }
       ok(false, "Worker had an error: " + event.message);
       worker.terminate();
       nextTestHarnessStep();
@@ -147,6 +154,10 @@ function testHarnessSteps() {
           nextTestHarnessStep();
           break;
 
+        case "expectUncaughtException":
+          worker._expectingUncaughtException = message.expecting;
+          break;
+
         case "clearAllDatabases":
           clearAllDatabases(function(){
             worker.postMessage({ op: "clearAllDatabasesDone" });
@@ -168,6 +179,11 @@ function testHarnessSteps() {
     URL.revokeObjectURL(workerScriptURL);
 
     yield undefined;
+
+    if (worker._expectingUncaughtException) {
+      ok(false, "expectUncaughtException was called but no uncaught " +
+                "exception was detected!");
+    }
 
     worker.terminate();
     worker = null;
@@ -243,6 +259,11 @@ function errorHandler(event)
 {
   ok(false, "indexedDB error, '" + event.target.error.name + "'");
   finishTest();
+}
+
+function expectUncaughtException(expecting)
+{
+  SimpleTest.expectUncaughtException(expecting);
 }
 
 function browserErrorHandler(event)
@@ -413,6 +434,10 @@ function workerScript() {
   };
 
   self.finishTest = function() {
+    if (self._expectingUncaughtException) {
+      self.ok(false, "expectUncaughtException was called but no uncaught "
+                     + "exception was detected!");
+    }
     self.postMessage({ op: "done" });
   };
 
@@ -508,6 +533,12 @@ function workerScript() {
     return buffer;
   };
 
+  self._expectingUncaughtException = false;
+  self.expectUncaughtException = function(_expecting_) {
+    self._expectingUncaughtException = !!_expecting_;
+    self.postMessage({ op: "expectUncaughtException", expecting: !!_expecting_ });
+  };
+
   self._clearAllDatabasesCallback = undefined;
   self.clearAllDatabases = function(_callback_) {
     self._clearAllDatabasesCallback = _callback_;
@@ -515,6 +546,12 @@ function workerScript() {
   }
 
   self.onerror = function(_message_, _file_, _line_) {
+    if (self._expectingUncaughtException) {
+      self._expectingUncaughtException = false;
+      ok(true, "Worker: expected exception [" + _file_ + ":" + _line_ + "]: '" +
+         _message_ + "'");
+      return;
+    }
     ok(false,
        "Worker: uncaught exception [" + _file_ + ":" + _line_ + "]: '" +
          _message_ + "'");
